@@ -11,40 +11,39 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.11",
-    "@angular/cdk": "^7.3.6",
-    "@angular/common": "^7.2.11",
-    "@angular/compiler": "^7.2.11",
-    "@angular/core": "^7.2.11",
-    "@angular/flex-layout": "^7.0.0-beta.24",
-    "@angular/forms": "^7.2.11",
-    "@angular/http": "^7.2.11",
-    "@angular/material": "^7.3.6",
-    "@angular/platform-browser": "^7.2.11",
-    "@angular/platform-browser-dynamic": "^7.2.11",
-    "@angular/router": "^7.2.11",
-    "@ngrx/store": "^7.4.0",
+    "@angular/animations": "^8.0.0",
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/flex-layout": "^8.0.0-beta.26",
+    "@angular/forms": "^8.0.0",
+    "@angular/material": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
+    "@ngrx/store": "^8.0.0",
     "@nguniversal/express-engine": "^7.1.1",
     "@nguniversal/module-map-ngfactory-loader": "^7.1.1",
-    "core-js": "^2.5.4",
+    "core-js": "^3.1.3",
     "express": "^4.17.1",
     "font-awesome": "4.7.0",
-    "npm-check-updates": "^3.1.4",
-    "ol": "^5.3.1",
-    "rxjs": "^6.4.0",
-    "systemjs": "^0.21.3",
-    "zone.js": "^0.9.0"
+    "npm-check-updates": "^3.1.10",
+    "ol": "^5.3.3",
+    "rxjs": "^6.5.2",
+    "systemjs": "^0.21.6",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular-devkit/build-ng-packagr": "~0.13.0",
-    "@angular/cli": "~7.3.7",
-    "@angular/compiler-cli": "^7.2.11",
-    "@angular/language-service": "^7.2.11",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular-devkit/build-ng-packagr": "~0.800.0",
+    "@angular/cli": "~8.0.2",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "^8.0.0",
     "@types/jasmine": "^3.3.12",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^11.13.0",
-    "codelyzer": "~5.0.0",
+    "codelyzer": "^5.0.1",
     "cross-env": "^5.1",
     "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
@@ -53,7 +52,7 @@
     "karma-coverage-istanbul-reporter": "^2.0.5",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.0",
-    "ng-packagr": "^5.2.0",
+    "ng-packagr": "^5.1.0",
     "prettier": "^1.17.1",
     "protractor": "^5.4.2",
     "ts-loader": "^6.0.1",
@@ -61,7 +60,7 @@
     "tsickle": "^0.35.0",
     "tslib": "^1.9.3",
     "tslint": "~5.15.0",
-    "typescript": "~3.2.2",
-    "webpack-cli": "^3.3.2"
+    "typescript": "~3.4.5",
+    "webpack-cli": "^3.3.3"
   }
 }
